+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - (Documentation) Improve Charts and Upgrade Documentation
 - (Feature) AutoDiscover Operator Access
 - (Feature) (Platform) Adjust Gateway timeouts
+- (Feature) (Platform) Docs and Installer improvements
 
 ## [1.2.47](https://github.com/arangodb/kube-arangodb/tree/1.2.47) (2025-03-28)
 - (Bugfix) Use Profile Annotations